Uploaded code for miniproject
We ran into a lot of technical difficulties, hence there are a lot of commented out sections. Moreover, the performance of the test.py script has been severely reduced due to me having to use HD video captured with my phone to test out the algorithm since I couldn't get the camera to run on startup so I could leave it recording over the St. Mary's bridge and use that video for the classifier. Most of the problems we got came from dealing with different video codecs and encapsulations and conversions and other related crap.